requirements.txt
openai>=1.30.0
pydantic>=2.6.0
python-dotenv>=1.0.0
streamlit>=1.35.0
numpy>=1.26.0